refactor(web-app): refactor files and directories
set file names to pascal case and push const, types and schemas directories to model directory
