+++ apps/web-app/src/models/types/index.ts
@@ -1,0 +1,1 @@
+export * from './formData.type'
