+++ apps/web-app/src/pages/Register/Register.tsx
@@ -3,10 +3,7 @@ import { GlobalLayout } from '@/layouts'
 
 import { RegisterFooter, RegisterHeader } from './components/'
 import { metadata } from './config'
-import { inputsData } from './consts'
-import { registerInputsSchema } from './schemas/'
-
-
+import { inputsData, registerInputsSchema } from './models'
 
 const Register = () => {
   return (
