+++ apps/web-app/src/pages/Register/models/consts/index.ts
