+++ apps/web-app/src/pages/Register/models/index.ts
@@ -1,0 +1,3 @@
+export * from './consts'
+export * from './schemas'
+
