+++ apps/web-app/src/pages/Register/models/schemas/registerInputs.schema.ts
@@ -1,6 +1,6 @@
+import { inputsData } from '@/pages/Register/models'
 import { getDynamicSchemas } from '@/utils'
 import { z as zValidator } from 'zod'
-import { inputsData } from '../consts'
 
 const registerInputs = getDynamicSchemas(inputsData)
