--- apps/web-app/src/routing/PublicRoutes.tsx
@@ -1,1 +1,0 @@
-sladkfjaslñfjadslfk
