+++ apps/web-app/src/routing/index.ts
@@ -1,4 +1,4 @@
-export * from './PrivateRoutes'
-export * from './Routes'
 export * from './Routing'
+export * from './guards/auth.guard'
+export * from './routes.model'
